# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trusted Neighbourhood Network API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/communities": {
      "get": {
        "tags": [
          "communities"
        ],
        "summary": "List Communities",
        "operationId": "list_communities_api_v1_communities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommunityOut"
                  },
                  "type": "array",
                  "title": "Response List Communities Api V1 Communities Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "communities"
        ],
        "summary": "Create Community",
        "operationId": "create_community_api_v1_communities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/communities/{id}": {
      "get": {
        "tags": [
          "communities"
        ],
        "summary": "Get Community",
        "operationId": "get_community_api_v1_communities__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "communities"
        ],
        "summary": "Update Community",
        "operationId": "update_community_api_v1_communities__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communities/{id}/units": {
      "get": {
        "tags": [
          "communities"
        ],
        "summary": "List Units",
        "operationId": "list_units_api_v1_communities__id__units_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnitOut"
                  },
                  "title": "Response List Units Api V1 Communities  Id  Units Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communities/{community_id}/units": {
      "post": {
        "tags": [
          "communities"
        ],
        "summary": "Add Units",
        "operationId": "add_units_api_v1_communities__community_id__units_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Community Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitBulkIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnitOut"
                  },
                  "title": "Response Add Units Api V1 Communities  Community Id  Units Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communities/{community_id}/members": {
      "post": {
        "tags": [
          "communities"
        ],
        "summary": "Request Membership",
        "operationId": "request_membership_api_v1_communities__community_id__members_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Community Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRequestIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "communities"
        ],
        "summary": "List Members",
        "operationId": "list_members_api_v1_communities__community_id__members_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Community Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MembershipOut"
                  },
                  "title": "Response List Members Api V1 Communities  Community Id  Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communities/{community_id}/members/{membership_id}": {
      "patch": {
        "tags": [
          "communities"
        ],
        "summary": "Update Member",
        "operationId": "update_member_api_v1_communities__community_id__members__membership_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Community Id"
            }
          },
          {
            "name": "membership_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Membership Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communities/{community_id}/members/import": {
      "post": {
        "tags": [
          "communities"
        ],
        "summary": "Import Residents",
        "operationId": "import_residents_api_v1_communities__community_id__members_import_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Community Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_residents_api_v1_communities__community_id__members_import_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_import_residents_api_v1_communities__community_id__members_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_residents_api_v1_communities__community_id__members_import_post"
      },
      "CommunityIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CommunityIn"
      },
      "CommunityOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "status",
          "settings",
          "created_at"
        ],
        "title": "CommunityOut"
      },
      "CommunityRole": {
        "type": "string",
        "enum": [
          "resident",
          "committee_member",
          "community_admin"
        ],
        "title": "CommunityRole"
      },
      "CommunityUpdateIn": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings"
          }
        },
        "type": "object",
        "title": "CommunityUpdateIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportRowResult": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "mobile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobile"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "row",
          "outcome"
        ],
        "title": "ImportRowResult"
      },
      "ImportSummary": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "existing": {
            "type": "integer",
            "title": "Existing"
          },
          "errors": {
            "type": "integer",
            "title": "Errors"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ImportRowResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "total",
          "created",
          "existing",
          "errors",
          "results"
        ],
        "title": "ImportSummary"
      },
      "JoinRequestIn": {
        "properties": {
          "unit_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Id"
          },
          "household_relationship": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Household Relationship"
          }
        },
        "type": "object",
        "title": "JoinRequestIn"
      },
      "MembershipOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "community_id": {
            "type": "string",
            "format": "uuid",
            "title": "Community Id"
          },
          "unit_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "household_relationship": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Household Relationship"
          },
          "capabilities": {
            "items": {},
            "type": "array",
            "title": "Capabilities"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "community_id",
          "unit_id",
          "role",
          "status",
          "verification_status",
          "household_relationship",
          "capabilities",
          "created_at"
        ],
        "title": "MembershipOut"
      },
      "MembershipStatus": {
        "type": "string",
        "enum": [
          "pending",
          "active",
          "suspended",
          "rejected"
        ],
        "title": "MembershipStatus"
      },
      "MembershipUpdateIn": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MembershipStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CommunityRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "unit_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Id"
          },
          "capabilities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capabilities"
          }
        },
        "type": "object",
        "title": "MembershipUpdateIn"
      },
...
      "UnitBulkIn": {
        "properties": {
          "units": {
            "items": {
              "$ref": "#/components/schemas/UnitIn"
            },
            "type": "array",
            "maxItems": 2000,
            "minItems": 1,
            "title": "Units"
          }
        },
        "type": "object",
        "required": [
          "units"
        ],
        "title": "UnitBulkIn"
      },
      "UnitIn": {
        "properties": {
          "tower": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Tower"
          },
          "unit_number": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Unit Number"
          }
        },
        "type": "object",
        "required": [
          "tower",
          "unit_number"
        ],
        "title": "UnitIn"
      },
      "UnitOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tower": {
            "type": "string",
            "title": "Tower"
          },
          "unit_number": {
            "type": "string",
            "title": "Unit Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tower",
          "unit_number"
        ],
        "title": "UnitOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}